+++ supvisors/application.py
@@ -17,30 +17,35 @@
 # limitations under the License.
 # ======================================================================
 
-from typing import Dict, List, Sequence
+from typing import Any, Dict, List, Sequence
 
 from supervisor.loggers import Logger
-from supervisor.states import *
+from supervisor.states import ProcessStates, STOPPED_STATES
 
 from .process import ProcessStatus
-from .ttypes import ApplicationStates, Payload, StartingStrategies, StartingFailureStrategies, RunningFailureStrategies
+from .ttypes import (ApplicationStates, NameList, Payload, StartingStrategies,
+                     StartingFailureStrategies, RunningFailureStrategies)
 
 
 class ApplicationRules(object):
     """ Definition of the rules for starting an application, iaw rules file.
 
     Attributes are:
-        - default: True if not overwritten by rules file
-        - start_sequence: defines the order of this application when starting all the applications in the DEPLOYMENT state (0 means: no automatic start),
-        - stop_sequence: defines the order of this application when stopping all the applications (0 means: immediate stop),
-        - starting_strategy: defines the strategy to apply when choosing the node where the process shall be started during the starting of the application,
-        - starting_failure_strategy: defines the strategy to apply when a required process cannot be started during the starting of the application,
+        - managed: set to True when application rules are found from the rules file;
+        - distributed: set to False if all processes must be running on the same node;
+        - node_names: the nodes where the application can be started (all by default) if not distributed,
+        - start_sequence: defines the order of this application when starting all the applications in the DEPLOYMENT state (0 means: no automatic start);
+        - stop_sequence: defines the order of this application when stopping all the applications (0 means: immediate stop);
+        - starting_strategy: defines the strategy to apply when choosing the node where the process shall be started during the starting of the application;
+        - starting_failure_strategy: defines the strategy to apply when a required process cannot be started during the starting of the application;
         - running_failure_strategy: defines the default strategy to apply when a required process crashes when the application is running.
     """
 
     def __init__(self) -> None:
         """ Initialization of the attributes. """
         self.managed: bool = False
+        self.distributed: bool = True
+        self.node_names: NameList = ['*']
         self.start_sequence: int = 0
         self.stop_sequence: int = 0
         self.starting_strategy: StartingStrategies = StartingStrategies.CONFIG
@@ -52,23 +57,29 @@ class ApplicationRules(object):
 
         :return: the printable application rules
         """
-        return 'managed={} start_sequence={} stop_sequence={} starting_strategy={}'\
+        return 'managed={} distributed={} node_names={} start_sequence={} stop_sequence={} starting_strategy={}'\
                ' starting_failure_strategy={} running_failure_strategy={}' \
-            .format(self.managed, self.start_sequence, self.stop_sequence, self.starting_strategy.name,
-                    self.starting_failure_strategy.name, self.running_failure_strategy.name)
+            .format(self.managed, self.distributed, self.node_names, self.start_sequence, self.stop_sequence,
+                    self.starting_strategy.name, self.starting_failure_strategy.name,
+                    self.running_failure_strategy.name)
 
     # serialization
     def serial(self) -> Payload:
         """ Get a serializable form of the application rules.
+        Do not send not applicable information.
 
         :return: the application rules in a dictionary
         """
-        return {'managed': self.managed,
-                'start_sequence': self.start_sequence,
-                'stop_sequence': self.stop_sequence,
-                'starting_strategy': self.starting_strategy.name,
-                'starting_failure_strategy': self.starting_failure_strategy.name,
-                'running_failure_strategy': self.running_failure_strategy.name}
+        if self.managed:
+            payload = {'managed': True, 'distributed': self.distributed,
+                       'start_sequence': self.start_sequence, 'stop_sequence': self.stop_sequence,
+                       'starting_strategy': self.starting_strategy.name,
+                       'starting_failure_strategy': self.starting_failure_strategy.name,
+                       'running_failure_strategy': self.running_failure_strategy.name}
+            if not self.distributed:
+                payload['addresses'] = self.node_names
+            return payload
+        return {'managed': False}
 
 
 class ApplicationStatus(object):
@@ -92,23 +103,24 @@ class ApplicationStatus(object):
     PrintableApplicationSequence = Dict[int, Sequence[str]]
     ProcessMap = Dict[str, ProcessStatus]
 
-    def __init__(self, application_name: str, rules: ApplicationRules, logger: Logger) -> None:
+    def __init__(self, application_name: str, rules: ApplicationRules, supvisors: Any) -> None:
         """ Initialization of the attributes.
 
         :param application_name: the name of the application
         :param rules: the rules applicable to the application
-        :param logger: the common logger used throughout Supvisors
+        :param supvisors: the global Supvisors structure
         """
         # keep reference to common logger
-        self.logger = logger
+        self.supvisors = supvisors
+        self.logger: Logger = supvisors.logger
         # information part
-        self.application_name = application_name
-        self._state = ApplicationStates.STOPPED
-        self.major_failure = False
-        self.minor_failure = False
+        self.application_name: str = application_name
+        self._state: ApplicationStates = ApplicationStates.STOPPED
+        self.major_failure: bool = False
+        self.minor_failure: bool = False
         # process part
         self.processes: ApplicationStatus.ProcessMap = {}
-        self.rules = rules
+        self.rules: ApplicationRules = rules
         self.start_sequence: ApplicationStatus.ApplicationSequence = {}
         self.stop_sequence: ApplicationStatus.ApplicationSequence = {}
 
@@ -197,6 +209,24 @@ class ApplicationStatus(object):
         """
         self.processes[process.process_name] = process
 
+    def possible_nodes(self) -> NameList:
+        """ Return the list of nodes where the application could be started.
+        To achieve that, two conditions:
+            - the Supervisor node must know all the application programs;
+            - the node must be declared in the applicable nodes in the rules file.
+
+        :return: the list of nodes where the program could be started
+        """
+        node_names = self.rules.node_names
+        if '*' in self.rules.node_names:
+            node_names = self.supvisors.address_mapper.node_names
+        # get the nodes common to all application processes
+        actual_nodes = [set(process.info_map.keys()) for process in self.processes.values()]
+        if actual_nodes:
+            actual_nodes = actual_nodes[0].intersection(*actual_nodes)
+        # intersect with rules
+        return [node_name for node_name in node_names if node_name in actual_nodes]
+
     @staticmethod
     def printable_sequence(application_sequence: ApplicationSequence) -> PrintableApplicationSequence:
         """ Get printable application sequence for log traces.
@@ -230,6 +260,19 @@ class ApplicationStatus(object):
             self.logger.info('ApplicationStatus.update_sequences: application_name={}'
                              ' is not managed so sequences are not defined'. format(self.application_name))
 
+    def get_start_sequence_expected_load(self) -> int:
+        """ Return the sum of the expected loading of the processes in the starting sequence.
+        This is used only in the event where the application is not distributed and the whole application loading
+        has to be checked on a single node.
+
+        :return: the expected loading of the application.
+        """
+        # get the processes from the starting sequence
+        # remove the sequence 0 as it corresponds to processes that are not meant to be autostarted
+        sequenced_processes = [process for seq, sub_seq in self.start_sequence.items()
+                               for process in sub_seq if seq > 0]
+        return sum(process.rules.expected_load for process in sequenced_processes)
+
     def update_status(self) -> None:
         """ Update the state of the application iaw the state of its sequenced processes.
         An unmanaged application - or generally without starting sequence - is always STOPPED.
@@ -243,8 +286,8 @@ class ApplicationStatus(object):
         sequenced_processes = [process for sub_seq in self.start_sequence.values()
                                for process in sub_seq]
         if not sequenced_processes:
-            # TODO: check if it is relevant to evaluate the application state in this case,
-            #  knowing that this application could include multiple "conflicts" due to its unmanaged nature
+            # TODO: check if it is relevant to evaluate the application state in this case
+            #  keep excluding unmanaged applications
             self.logger.debug('ApplicationStatus.update_status: application_name={}'
                               ' is not managed so always STOPPED'. format(self.application_name))
             self.state = ApplicationStates.STOPPED