+++ supvisors/tests/configurations.py
@@ -121,6 +121,7 @@ InvalidXmlTest = b'''\
     </application>
 
     <application name="dummy_application_B">
+        <distributed>non</distributed>
         <start_sequence>1</start_sequence>
         <stop_sequence>4</stop_sequence>
         <starting_failure_strategy>STOP</starting_failure_strategy>
@@ -176,6 +177,8 @@ InvalidXmlTest = b'''\
     </application>
 
     <application name="dummy_application_C">
+        <distributed>false</distributed>
+        <addresses>192.256.16.10,*</addresses>
         <start_sequence>20</start_sequence>
         <stop_sequence>0</stop_sequence>
         <starting_failure_strategy>ABORT</starting_failure_strategy>
@@ -275,6 +278,7 @@ XmlTest = b'''\
     </application>
 
     <application name="dummy_application_B">
+        <distributed>false</distributed>
         <start_sequence>1</start_sequence>
         <stop_sequence>4</stop_sequence>
         <starting_strategy>CONFIG</starting_strategy>
@@ -315,6 +319,7 @@ XmlTest = b'''\
     </application>
 
     <application name="dummy_application_C">
+        <distributed>true</distributed>
         <start_sequence>20</start_sequence>
         <stop_sequence>0</stop_sequence>
         <starting_strategy>LOCAL</starting_strategy>
@@ -340,6 +345,8 @@ XmlTest = b'''\
     </application>
 
     <application name="dummy_application_D">
+        <distributed>false</distributed>
+        <addresses>10.0.0.1, 10.0.0.5</addresses>
         <start_sequence>-1</start_sequence>
         <stop_sequence>100</stop_sequence>
         <starting_strategy>LESS_LOADED</starting_strategy>