+++ supvisors/tests/test_sparser.py
@@ -30,14 +30,16 @@ from .configurations import InvalidXmlTest, XmlTest
 
 def assert_default_application_rules(rules):
     """ Check that rules contains default values. """
-    assert_application_rules(rules, False, 0, 0, StartingStrategies.CONFIG,
+    assert_application_rules(rules, False, True, ['*'], 0, 0, StartingStrategies.CONFIG,
                              StartingFailureStrategies.ABORT, RunningFailureStrategies.CONTINUE)
 
 
-def assert_application_rules(rules, managed, start, stop, starting_strategy,
+def assert_application_rules(rules, managed, distributed, node_names, start, stop, starting_strategy,
                              starting_failure_strategy, running_failure_strategy):
     """ Check the application rules. """
     assert rules.managed == managed
+    assert rules.distributed == distributed
+    assert rules.node_names == node_names
     assert rules.start_sequence == start
     assert rules.stop_sequence == stop
     assert rules.starting_strategy == starting_strategy
@@ -86,23 +88,23 @@ def check_valid(parser):
     assert_default_application_rules(rules)
     # check first application
     rules = load_application_rules(parser, 'dummy_application_A')
-    assert_application_rules(rules, True, 0, 0, StartingStrategies.CONFIG,
+    assert_application_rules(rules, True, True, ['*'], 0, 0, StartingStrategies.CONFIG,
                              StartingFailureStrategies.ABORT, RunningFailureStrategies.CONTINUE)
     # check second application
     rules = load_application_rules(parser, 'dummy_application_B')
-    assert_application_rules(rules, True, 1, 4, StartingStrategies.CONFIG,
+    assert_application_rules(rules, True, False, ['*'], 1, 4, StartingStrategies.CONFIG,
                              StartingFailureStrategies.STOP, RunningFailureStrategies.RESTART_PROCESS)
     # check third application
     rules = load_application_rules(parser, 'dummy_application_C')
-    assert_application_rules(rules, True, 20, 0, StartingStrategies.LOCAL,
+    assert_application_rules(rules, True, True, ['*'], 20, 0, StartingStrategies.LOCAL,
                              StartingFailureStrategies.ABORT, RunningFailureStrategies.STOP_APPLICATION)
     # check fourth application
     rules = load_application_rules(parser, 'dummy_application_D')
-    assert_application_rules(rules, True, 0, 100, StartingStrategies.LESS_LOADED,
+    assert_application_rules(rules, True, False, ['10.0.0.1', '10.0.0.5'], 0, 100, StartingStrategies.LESS_LOADED,
                              StartingFailureStrategies.CONTINUE, RunningFailureStrategies.RESTART_APPLICATION)
     # check loop application
     rules = load_application_rules(parser, 'dummy_application_E')
-    assert_application_rules(rules, True, 0, 0,  StartingStrategies.MOST_LOADED,
+    assert_application_rules(rules, True, True,  ['*'], 0, 0,  StartingStrategies.MOST_LOADED,
                              StartingFailureStrategies.ABORT, RunningFailureStrategies.CONTINUE)
     # check program from unknown application: all default
     rules = load_process_rules(parser, 'dummy_application_X', 'dummy_program_X0')
@@ -176,19 +178,19 @@ def check_invalid(parser):
     assert_default_application_rules(rules)
     # check first application
     rules = load_application_rules(parser, 'dummy_application_A')
-    assert_application_rules(rules, True, 0, 0, StartingStrategies.CONFIG,
+    assert_application_rules(rules, True, True, ['*'], 0, 0, StartingStrategies.CONFIG,
                              StartingFailureStrategies.ABORT, RunningFailureStrategies.CONTINUE)
     # check second application
     rules = load_application_rules(parser, 'dummy_application_B')
-    assert_application_rules(rules, True, 1, 4, StartingStrategies.CONFIG,
+    assert_application_rules(rules, True, True, ['*'], 1, 4, StartingStrategies.CONFIG,
                              StartingFailureStrategies.STOP, RunningFailureStrategies.RESTART_PROCESS)
     # check third application
     rules = load_application_rules(parser, 'dummy_application_C')
-    assert_application_rules(rules, True, 20, 0, StartingStrategies.CONFIG,
+    assert_application_rules(rules, True, False, ['*'], 20, 0, StartingStrategies.CONFIG,
                              StartingFailureStrategies.ABORT, RunningFailureStrategies.STOP_APPLICATION)
     # check fourth application
     rules = load_application_rules(parser, 'dummy_application_D')
-    assert_application_rules(rules, True, 0, 100, StartingStrategies.CONFIG,
+    assert_application_rules(rules, True, True, ['*'], 0, 100, StartingStrategies.CONFIG,
                              StartingFailureStrategies.CONTINUE, RunningFailureStrategies.RESTART_APPLICATION)
     # check program from unknown application: all default
     rules = load_process_rules(parser, 'dummy_application_X', 'dummy_program_X0')